Add functionality of filter heading
